Add floor map editing, image processing service, russian error localization and tests

+++ admin_service/src/main/java/com/hse/adminservice/operations/analytics/dto/UserAnalyticsResponse.java
@@ -14,8 +14,8 @@ public record UserAnalyticsResponse(
         Integer activeBookings,
         Integer unfinishedServiceRequests,
         Integer openPayRequests,
-        Integer totalBalance,
-        Integer monthlyIncome,
+        Long totalBalance,
+        Long monthlyIncome,
         Integer monthlyOccupancyPercent,
         List<AnalyticsMetricPointResponse> monthlyIncomeHistory,
         List<AnalyticsMetricPointResponse> occupancyHistory
