+++ admin_service/src/main/java/com/hse/adminservice/operations/queue/dto/PayRequestQueueItemResponse.java
@@ -10,7 +10,7 @@ public record PayRequestQueueItemResponse(
         Long membershipId,
         Long userId,
         String userName,
-        Integer amount,
+        Long amount,
         String status,
         String userComment,
         String adminComment,
