+++ admin_service/src/main/java/com/hse/adminservice/operations/users/dto/CoworkingUserReadModelResponse.java
@@ -9,7 +9,7 @@ public record CoworkingUserReadModelResponse(
         Long userId,
         String name,
         LocalDate registeredAt,
-        Integer balance,
+        Long balance,
         Integer totalBookings,
         Integer unfinishedBookings
 ) {
